requirements.txt
Django==2.2.4
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
django-appconf==1.0.3
django-image-cropping==1.2.0
easy-thumbnails==2.6
certifi==2019.6.16
chardet==3.0.4
django-embed-video==1.3
idna==2.8
requests==2.22.0
urllib3==1.25.3
